feat : 특정 버스 위치 가져오기 추가

+++ src/main/java/koreatechbus/domain/Bus.java
@@ -75,7 +75,7 @@ public class Bus {
             bookmarkNum = 0L;
     }
 
-    public void setPosition(Long latitude, Long longitude){
+    public void setPosition(Double latitude, Double longitude){
         this.latitude = latitude;
         this.longitude = longitude;
     }
